# Don't round off subseconds unless necessary